third_party/blink/web_tests/storage/indexeddb/queued-commands-expected.txt

Verify that queuing up several commands works (and they all fire).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "queued-commands.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
db.createObjectStore('storeName')
store.createIndex('indexName', 'x')
store.add({x: 'value', y: 'zzz'}, 'key')
store.add({x: 'value2', y: 'zzz2'}, 'key2')
store.put({x: 'valu2', y: 'zz2'}, 'ky2')
PASS 0 is 0
PASS 1 is 1
PASS 2 is 2
PASS successfullyParsed is true

TEST COMPLETE
